feat: move networks to dao
