--- apps/website/constants/networks.ts
@@ -1,13 +1,0 @@
-export const networks = [
-  'discord',
-  'email',
-  'github',
-  'medium',
-  'other',
-  'telegram',
-  'twitter',
-  'twitch',
-  'website',
-] as const;
-
-export type Network = typeof networks[number];
